requirements.txt
# ──────────────────────────────────────────────────────────────────────────────
# InstaFlow Automation Engine — Python Dependencies
# Pinned to stable ranges. Run: pip install -r requirements.txt
# ──────────────────────────────────────────────────────────────────────────────

# ── Core Framework ────────────────────────────────────────────────────────────
fastapi>=0.111.0,<0.120.0
uvicorn[standard]>=0.30.0,<0.35.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0

# ── Task Queue ────────────────────────────────────────────────────────────────
celery[redis]>=5.4.0,<6.0.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.30,<3.0.0
asyncpg>=0.29.0,<0.31.0          # Async PostgreSQL driver
psycopg2-binary>=2.9.0,<3.0.0   # Sync driver for Alembic migrations
alembic>=1.13.0,<2.0.0

# ── Redis ─────────────────────────────────────────────────────────────────────
redis[hiredis]>=5.0.0,<6.0.0

# ── Instagram ─────────────────────────────────────────────────────────────────
instagrapi>=2.1.0,<3.0.0

# ── Browser Automation ────────────────────────────────────────────────────────
playwright>=1.44.0,<2.0.0

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx>=0.27.0,<0.28.0

# ── Encryption ────────────────────────────────────────────────────────────────
cryptography>=42.0.0,<44.0.0

# ── Config / YAML ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.0,<7.0.0

# ── Logging ───────────────────────────────────────────────────────────────────
structlog>=24.1.0,<25.0.0

# ── Excel Export ──────────────────────────────────────────────────────────────
openpyxl>=3.1.0,<4.0.0

# ── Alerting ──────────────────────────────────────────────────────────────────
python-telegram-bot>=21.0,<22.0 

# ── Dev / Testing (not needed in production) ──────────────────────────────────
pytest>=8.2.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
pytest-mock>=3.14.0,<4.0.0
httpx>=0.27.0                    # also used for test client
